requirements.txt
streamlit
pinecone-client
sentence-transformers
transformers
torch